+++ include/graphics/UpperRenderer.h
@@ -6,8 +6,6 @@
 #include "graphics/GL.h"
 #include "threading/Lockable.h"
 
-#include "lib/Eigen.h"
-
 #include <atomic>
 #include <future>
 #include <unordered_set>
@@ -21,7 +19,7 @@ namespace Game3 {
 			int backbufferWidth = -1;
 			int backbufferHeight = -1;
 
-			Eigen::Vector2f center{0.f, 0.f};
+			Vector2d center{0, 0};
 			std::shared_ptr<Tileset> tileset{};
 
 			UpperRenderer();
